generator/requirements.txt
kafka-python==2.0.2
ujson==5.10.0

